refactor
Some cleaning and some refactoring
