--- BFF/src/main/kotlin/br/com/example/simple/controller/ScreenController.kt
@@ -1,13 +1,0 @@
-package br.com.example.simple.controller
-
-import br.com.example.simple.service.OnboardingScreenService
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class ScreenController(
-        private val onboardingScreenService: OnboardingScreenService
-){
-    @GetMapping("/onboarding")
-    fun getOmboardingScreen() = onboardingScreenService.createdOnboardingScreen()
-}
