+++ BFF/src/main/kotlin/br/com/example/simple/screen/OnboardingScreen.kt
@@ -1,4 +1,4 @@
-package br.com.example.simple.builder
+package br.com.example.simple.screen
 
 import br.com.zup.beagle.ext.applyFlex
 import br.com.zup.beagle.widget.core.AlignSelf
@@ -9,14 +9,12 @@ import br.com.zup.beagle.widget.layout.Screen
 import br.com.zup.beagle.widget.layout.ScreenBuilder
 import br.com.zup.beagle.widget.ui.Text
 
-class OnboardingScreenBuilder: ScreenBuilder {
-    override fun build(): Screen {
-        return Screen(
+class OnboardingScreen : ScreenBuilder {
+    override fun build() = Screen(
                 child = Container(
                         children = listOf(
                                 Text("Hello Beagle").applyFlex(flex = Flex(alignSelf = AlignSelf.CENTER))
                         )
                 ).applyFlex(flex = Flex(grow = 1.0,justifyContent = JustifyContent.CENTER))
         )
-    }
 }
