--- BFF/src/main/kotlin/br/com/example/simple/service/OnboardingScreenService.kt
@@ -1,9 +1,0 @@
-package br.com.example.simple.service
-
-import br.com.example.simple.builder.OnboardingScreenBuilder
-import org.springframework.stereotype.Service
-
-@Service
-class OnboardingScreenService {
-    fun createdOnboardingScreen() = OnboardingScreenBuilder().build()
-}
