+++ BFF/src/main/kotlin/br/com/example/simple/service/OnboardingService.kt
@@ -1,0 +1,9 @@
+package br.com.example.simple.service
+
+import br.com.example.simple.screen.OnboardingScreen
+import org.springframework.stereotype.Service
+
+@Service
+class OnboardingService {
+    fun getOnboardingScreen() = OnboardingScreen()
+}
