+++ BFF/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+spring.jackson.default-property-inclusion=non_null
