--- BFF/src/test/kotlin/br/com/example/simple/SimpleApplicationTests.kt
@@ -1,13 +1,0 @@
-package br.com.example.simple
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class SimpleApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
